backend/requirements.txt
django>=3.2,<5.0
psycopg2-binary
djangorestframework
django-cors-headers
gunicorn
whitenoise
dj-database-url
